+++ app/Models/Turno.php
@@ -2,9 +2,10 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Turno extends Model
 {
-    //
+    use HasFactory;
 }
