+++ resources/views/components/layout.blade.php
@@ -7,8 +7,8 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>{{ $metaTitle ?? 'Default title' }}</title>
     <meta name="description" content="{{ $metaDescription ?? 'Default description' }}" />
-    @vite(['resources/css/app.css', 'resources/js/app.js'])
-    {{-- @vite(['resources/css/app.css', 'resources/css/styleIndex.css', 'resources/js/app.js', 'resources/js/scriptIndex.js']) --}}
+    {{-- @vite(['resources/css/app.css', 'resources/js/app.js']) --}}
+        @vite(['resources/css/app.css', 'resources/css/styleIndex.css', 'resources/js/app.js', 'resources/js/scriptIndex.js'])
 
 </head>
 
@@ -21,6 +21,7 @@
     <main class="flex-1 p-4">
         {{ $slot }}
     </main>
+    
 </body>
 <footer class="py-10 px-4">
     <div class="mx-auto flex max-w6 flex-col items-center space-y-4 md:flex md:justify-between">
